Assignment 4
this is the fourth assignment in my programming 2 course. it builds on assignment 3. in this assignment we have done away with structs and are now using classes  only. the parent and chlid class concept has been introduced with employee and hourly employee. Other concepts intorduced in this assignment include copy constructors and assignment operator overloading  Our information is still being stored in a singly linked list. Also starting last assignment we started implementting a make file.